pyproject.toml
[tool.poetry]
name = "trailhub"
version = "0.1.0"
description = "Trailhub is a system for logging actions performed on your system"
authors = ["flavien-hugs <flavienhugs@pm.me>"]
packages = [{include = "src" }]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
fastapi = {version = "^0.112.2", extras = ["all"]}
beanie = "1.26.0"
uvloop = "^0.21.0"
pydantic-settings = "^2.6.1"
fastapi-pagination = "^0.12.32"
python-slugify = "^8.0.4"
httptools = "^0.6.4"
user-agents = "^2.2.0"
getmac = "^0.9.5"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
black = "^24.10.0"
flake8 = "^7.1.1"
bandit = "^1.8.0"
flake8-pyproject = "^1.2.3"

scalene = "^1.5.49"
[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
coverage = "^7.6.8"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.24.0"
mongomock-motor = "^0.0.34"
pytest-dotenv = "^0.5.2"
faker = "^33.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = 'src.run:app'

[tool.flake8]
max-line-length = 130
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    ".pytest_cache",
    ".tox",
    ".eggs"
]

[tool.black]
line-length = 130
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \.docker
    | \*.egg-info
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.pytest.ini_options]
env_files = "tests/.test.env"
env_override_existing_values = 1
capture = "no"
log-cli-level = "INFO"
python_files = "test_*.py"
asyncio_mode = "auto"
filterwarnings = [
  "ignore",
  "ignore:.*U.*mode is deprecated:DeprecationWarning"
]
